+++ donkeycar/parts/actuator.py
@@ -542,4 +542,7 @@ class Mini_HBridge_DC_Motor_PWM(object):
         import RPi.GPIO as GPIO
         self.pwm_f.stop()
         self.pwm_b.stop()
-        GPIO.cleanup()
+        try:
+            GPIO.cleanup()
+        except:
+            pass
