+++ donkeycar/templates/donkey2.py
@@ -1,15 +1,15 @@
 #!/usr/bin/env python3
 """
-Scripts to drive a donkey 2 car and train a model for it. 
+Scripts to drive a donkey 2 car
 
 Usage:
-    manage.py (drive) [--model=<model>] [--js]
-    manage.py (train) [--tub=<tub1,tub2,..tubn>]  (--model=<model>) [--no_cache]
+    manage.py (drive) [--model=<model>] [--js] [--type=(linear|categorical|rnn|imu|behavior|3d)] [--camera=(single|stereo)]
+    manage.py (train) [--tub=<tub1,tub2,..tubn>] (--model=<model>) [--transfer=<model>] [--type=(linear|categorical|rnn|imu|behavior|3d)] [--continuous] [--aug]
+
 
 Options:
-    -h --help        Show this screen.
-    --tub TUBPATHS   List of paths to tubs. Comma separated. Use quotes to use wildcards. ie "~/tubs/*"
-    --js             Use physical joystick.
+    -h --help     Show this screen.
+    --js          Use physical joystick.
 """
 import os
 from docopt import docopt
@@ -17,15 +17,15 @@ from docopt import docopt
 import donkeycar as dk
 
 #import parts
-from donkeycar.parts.camera import PiCamera
 from donkeycar.parts.transform import Lambda
-from donkeycar.parts.actuator import PCA9685, PWMSteering, PWMThrottle
 from donkeycar.parts.datastore import TubHandler, TubGroup
 from donkeycar.parts.controller import LocalWebController, JoystickController
+from donkeycar.parts.imu import Mpu6050
+import numpy as np
+from donkeycar.parts.throttle_filter import ThrottleFilter
+from donkeycar.parts.behavior import BehaviorPart
 
-
-
-def drive(cfg, model_path=None, use_joystick=False):
+def drive(cfg, model_path=None, use_joystick=False, model_type=None, camera_type='single'):
     '''
     Construct a working robotic vehicle from many parts.
     Each part runs as a job in the Vehicle loop, calling either
@@ -36,15 +36,76 @@ def drive(cfg, model_path=None, use_joystick=False):
     to parts requesting the same named input.
     '''
 
+    if model_type is None:
+        model_type = "categorical"
+    
     #Initialize car
     V = dk.vehicle.Vehicle()
-    cam = PiCamera(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H)
-    V.add(cam, outputs=['cam/image_array'], threaded=True)
-    
+
+    if camera_type == "stereo":
+
+        if cfg.CAMERA_TYPE == "WEBCAM":
+            from donkeycar.parts.camera import Webcam            
+
+            camA = Webcam(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H, iCam = 0)
+            camB = Webcam(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H, iCam = 1)
+
+        elif cfg.CAMERA_TYPE == "CVCAM":
+            from donkeycar.parts.cv import CvCam
+
+            camA = CvCam(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H, iCam = 0)
+            camB = CvCam(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H, iCam = 1)
+        else:
+            raise(Exception("Unsupported camera type: %s" % cfg.CAMERA_TYPE))
+
+        V.add(camA, outputs=['cam/image_array_a'], threaded=True)
+        V.add(camB, outputs=['cam/image_array_b'], threaded=True)
+
+        def stereo_pair(image_a, image_b):
+            '''
+            This will take the two images and combine them into a single image
+            One in red, the other in green, and diff in blue channel.
+            '''
+            if image_a is not None and image_b is not None:
+                width, height, _ = image_a.shape
+                grey_a = dk.utils.rgb2gray(image_a)
+                grey_b = dk.utils.rgb2gray(image_b)
+                grey_c = grey_a - grey_b
+                
+                stereo_image = np.zeros([width, height, 3], dtype=np.dtype('B'))
+                stereo_image[...,0] = np.reshape(grey_a, (width, height))
+                stereo_image[...,1] = np.reshape(grey_b, (width, height))
+                stereo_image[...,2] = np.reshape(grey_c, (width, height))
+            else:
+                stereo_image = []
+
+            return np.array(stereo_image)
+
+        image_sterero_pair_part = Lambda(stereo_pair)
+        V.add(image_sterero_pair_part, inputs=['cam/image_array_a', 'cam/image_array_b'], 
+            outputs=['cam/image_array'])
+
+    else:
+        
+        print("cfg.CAMERA_TYPE", cfg.CAMERA_TYPE)
+        if cfg.CAMERA_TYPE == "PICAM":
+            from donkeycar.parts.camera import PiCamera
+            cam = PiCamera(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H)
+        elif cfg.CAMERA_TYPE == "WEBCAM":
+            from donkeycar.parts.camera import Webcam
+            cam = Webcam(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H)
+        elif cfg.CAMERA_TYPE == "CVCAM":
+            from donkeycar.parts.cv import CvCam
+            cam = CvCam(image_w=cfg.IMAGE_W, image_h=cfg.IMAGE_H)
+        else:
+            raise(Exception("Unkown camera type: %s" % cfg.CAMERA_TYPE))
+            
+        V.add(cam, outputs=['cam/image_array'], threaded=True)
+        
     if use_joystick or cfg.USE_JOYSTICK_AS_DEFAULT:
         #modify max_throttle closer to 1.0 to have more power
         #modify steering_scale lower than 1.0 to have less responsive steering
-        ctr = JoystickController(max_throttle=cfg.JOYSTICK_MAX_THROTTLE,
+        ctr = JoystickController(throttle_scale=cfg.JOYSTICK_MAX_THROTTLE,
                                  steering_scale=cfg.JOYSTICK_STEERING_SCALE,
                                  auto_record_on_throttle=cfg.AUTO_RECORD_ON_THROTTLE,
                                  controller_type=cfg.CONTROLLER_TYPE)
@@ -65,6 +126,10 @@ def drive(cfg, model_path=None, use_joystick=False):
           inputs=['cam/image_array'],
           outputs=['user/angle', 'user/throttle', 'user/mode', 'recording'],
           threaded=True)
+
+    #this throttle filter will allow one tap back for esc reverse
+    th_filter = ThrottleFilter()
+    V.add(th_filter, inputs=['user/throttle'], outputs=['user/throttle'])
     
     #See if we should even run the pilot module. 
     #This is only needed because the part run_condition only accepts boolean
@@ -77,16 +142,74 @@ def drive(cfg, model_path=None, use_joystick=False):
     pilot_condition_part = Lambda(pilot_condition)
     V.add(pilot_condition_part, inputs=['user/mode'], outputs=['run_pilot'])
     
-    #Run the pilot if the mode is not user.
+    
+    def led_cond(mode, recording, num_records, behavior_state):
+        #returns a blink rate. 0 for off. -1 for on. positive for rate.
+        
+        if num_records is not None and num_records % 10 == 0:
+            if led_cond.last_num_rec_print != num_records:
+                print("recorded", num_records, "records")
+                led_cond.last_num_rec_print = num_records
+
+        if behavior_state is not None and model_type == 'behavior':
+            r, g, b = cfg.BEHAVIOR_LED_COLORS[behavior_state]
+            led.set_rgb(r, g, b)
+            return -1 #solid on
+
+        if recording:
+            return -1 #solid on
+        elif mode == 'user':
+            return 1
+        elif mode == 'local_angle':
+            return 0.5
+        elif mode == 'local':
+            return 0.1
+        return 0 
+
+    led_cond.last_num_rec_print = 0
+
+    led_cond_part = Lambda(led_cond)
+    V.add(led_cond_part, inputs=['user/mode', 'recording', "tub/num_records", 'behavior/state'], outputs=['led/blink_rate'])
+
+    if cfg.HAVE_RGB_LED:
+        from donkeycar.parts.led_status import RGB_LED
+        led = RGB_LED(cfg.LED_PIN_R, cfg.LED_PIN_G, cfg.LED_PIN_B, cfg.LED_INVERT)
+        led.set_rgb(cfg.LED_R, cfg.LED_G, cfg.LED_B)
+        V.add(led, inputs=['led/blink_rate'])
+
+    #IMU
+    if cfg.HAVE_IMU:
+        imu = Mpu6050()
+        V.add(imu, outputs=['imu/acl_x', 'imu/acl_y', 'imu/acl_z',
+            'imu/gyr_x', 'imu/gyr_y', 'imu/gyr_z'], threaded=True)
+
+    #Behavioral state
+    if model_type == "behavior":
+        bh = BehaviorPart(cfg.BEHAVIOR_LIST)
+        V.add(bh, outputs=['behavior/state', 'behavior/label', "behavior/one_hot_state_array"])
+        try:
+            ctr.set_button_down_trigger('L1', bh.increment_state)
+        except:
+            pass
+
+        inputs = ['cam/image_array', "behavior/one_hot_state_array"]  
+    #IMU
+    elif model_type == "imu":
+        assert(cfg.HAVE_IMU)
+        #Run the pilot if the mode is not user.
+        inputs=['cam/image_array',
+            'imu/acl_x', 'imu/acl_y', 'imu/acl_z',
+            'imu/gyr_x', 'imu/gyr_y', 'imu/gyr_z']
+    else:
+        inputs=['cam/image_array']
+
     if model_path:
-        from donkeycar.parts.keras import KerasCategorical
-        kl = KerasCategorical()
+        kl = dk.utils.get_model_by_type(model_type, cfg)
         kl.load(model_path)
     
-        V.add(kl, inputs=['cam/image_array'], 
+        V.add(kl, inputs=inputs, 
             outputs=['pilot/angle', 'pilot/throttle'],
-            run_condition='run_pilot')
-    
+            run_condition='run_pilot')            
     
     #Choose what inputs should change the car.
     def drive_mode(mode, 
@@ -107,74 +230,86 @@ def drive(cfg, model_path=None, use_joystick=False):
                   'pilot/angle', 'pilot/throttle'], 
           outputs=['angle', 'throttle'])
     
+    #Drive train setup
     
-    steering_controller = PCA9685(cfg.STEERING_CHANNEL)
-    steering = PWMSteering(controller=steering_controller,
-                                    left_pulse=cfg.STEERING_LEFT_PWM, 
-                                    right_pulse=cfg.STEERING_RIGHT_PWM)
+    if cfg.DRIVE_TRAIN_TYPE == "SERVO_ESC":
+        from donkeycar.parts.actuator import PCA9685, PWMSteering, PWMThrottle
+
+        steering_controller = PCA9685(cfg.STEERING_CHANNEL, cfg.PCA9685_I2C_ADDR, busnum=cfg.PCA9685_I2C_BUSNUM)
+        steering = PWMSteering(controller=steering_controller,
+                                        left_pulse=cfg.STEERING_LEFT_PWM, 
+                                        right_pulse=cfg.STEERING_RIGHT_PWM)
+        
+        throttle_controller = PCA9685(cfg.THROTTLE_CHANNEL, cfg.PCA9685_I2C_ADDR, busnum=cfg.PCA9685_I2C_BUSNUM)
+        throttle = PWMThrottle(controller=throttle_controller,
+                                        max_pulse=cfg.THROTTLE_FORWARD_PWM,
+                                        zero_pulse=cfg.THROTTLE_STOPPED_PWM, 
+                                        min_pulse=cfg.THROTTLE_REVERSE_PWM)
+
+        V.add(steering, inputs=['angle'])
+        V.add(throttle, inputs=['throttle'])
     
-    throttle_controller = PCA9685(cfg.THROTTLE_CHANNEL)
-    throttle = PWMThrottle(controller=throttle_controller,
-                                    max_pulse=cfg.THROTTLE_FORWARD_PWM,
-                                    zero_pulse=cfg.THROTTLE_STOPPED_PWM, 
-                                    min_pulse=cfg.THROTTLE_REVERSE_PWM)
+
+    elif cfg.DRIVE_TRAIN_TYPE == "DC_STEER_THROTTLE":
+        from donkeycar.parts.actuator import Mini_HBridge_DC_Motor_PWM
+        
+        steering = Mini_HBridge_DC_Motor_PWM(cfg.HBRIDGE_PIN_LEFT, cfg.HBRIDGE_PIN_RIGHT)
+        throttle = Mini_HBridge_DC_Motor_PWM(cfg.HBRIDGE_PIN_FWD, cfg.HBRIDGE_PIN_BWD)
+
+        V.add(steering, inputs=['angle'])
+        V.add(throttle, inputs=['throttle'])
     
-    V.add(steering, inputs=['angle'])
-    V.add(throttle, inputs=['throttle'])
+
+    elif cfg.DRIVE_TRAIN_TYPE == "DC_TWO_WHEEL":
+        from donkeycar.parts.actuator import TwoWheelSteeringThrottle, Mini_HBridge_DC_Motor_PWM
+
+        left_motor = Mini_HBridge_DC_Motor_PWM(cfg.HBRIDGE_PIN_LEFT_FWD, cfg.HBRIDGE_PIN_LEFT_BWD)
+        right_motor = Mini_HBridge_DC_Motor_PWM(cfg.HBRIDGE_PIN_RIGHT_FWD, cfg.HBRIDGE_PIN_RIGHT_BWD)
+        two_wheel_control = TwoWheelSteeringThrottle()
+
+        V.add(two_wheel_control, 
+                inputs=['throttle', 'angle'],
+                outputs=['left_motor_speed', 'right_motor_speed'])
+
+        V.add(left_motor, inputs=['left_motor_speed'])
+        V.add(right_motor, inputs=['right_motor_speed'])
+    
     
     #add tub to save data
-    inputs=['cam/image_array', 'user/angle', 'user/throttle', 'user/mode']
-    types=['image_array', 'float', 'float',  'str']
+
+    inputs=['cam/image_array',
+            'user/angle', 'user/throttle', 
+            'user/mode']
+
+    types=['image_array',
+           'float', 'float',  
+           'str']
+
+    if cfg.TRAIN_BEHAVIORS:
+        inputs += ['behavior/state', 'behavior/label', "behavior/one_hot_state_array"]
+        types += ['int', 'str', 'vector']
+    
+    if cfg.HAVE_IMU:
+        inputs += ['imu/acl_x', 'imu/acl_y', 'imu/acl_z',
+            'imu/gyr_x', 'imu/gyr_y', 'imu/gyr_z']
+
+        types +=['float', 'float', 'float',
+           'float', 'float', 'float']
     
     th = TubHandler(path=cfg.DATA_PATH)
     tub = th.new_tub_writer(inputs=inputs, types=types)
-    V.add(tub, inputs=inputs, outputs=['tub/num_records'], run_condition='recording')
-    
+    V.add(tub, inputs=inputs, outputs=["tub/num_records"], run_condition='recording')
+
+    if type(ctr) is LocalWebController:
+        print("You can now go to <your pi ip address>:8887 to drive your car.")
+    elif type(ctr) is JoystickController:
+        print("You can now move your joystick to drive your car.")
+        #tell the controller about the tub        
+        ctr.set_tub(tub)
+
     #run the vehicle for 20 seconds
     V.start(rate_hz=cfg.DRIVE_LOOP_HZ, 
             max_loop_count=cfg.MAX_LOOPS)
-    
-    print("You can now go to <your pi ip address>:8887 to drive your car.")
-
-
-def train(cfg, tub_names, model_name):
-    '''
-    use the specified data in tub_names to train an artifical neural network
-    saves the output trained model as model_name
-    '''
-    X_keys = ['cam/image_array']
-    y_keys = ['user/angle', 'user/throttle']
-
-    def rt(record):
-        record['user/angle'] = dk.utils.linear_bin(record['user/angle'])
-        return record
-
-    kl = KerasCategorical()
-    print('tub_names', tub_names)
-    if not tub_names:
-        tub_names = os.path.join(cfg.DATA_PATH, '*')
-    tubgroup = TubGroup(tub_names)
-    train_gen, val_gen = tubgroup.get_train_val_gen(X_keys, y_keys, record_transform=rt,
-                                                    batch_size=cfg.BATCH_SIZE,
-                                                    train_frac=cfg.TRAIN_TEST_SPLIT)
-
-    model_path = os.path.expanduser(model_name)
-
-    total_records = len(tubgroup.df)
-    total_train = int(total_records * cfg.TRAIN_TEST_SPLIT)
-    total_val = total_records - total_train
-    print('train: %d, validation: %d' % (total_train, total_val))
-    steps_per_epoch = total_train // cfg.BATCH_SIZE
-    print('steps_per_epoch', steps_per_epoch)
-
-    kl.train(train_gen,
-             val_gen,
-             saved_model_path=model_path,
-             steps=steps_per_epoch,
-             train_split=cfg.TRAIN_TEST_SPLIT)
-
-
-
 
 
 if __name__ == '__main__':
@@ -182,15 +317,20 @@ if __name__ == '__main__':
     cfg = dk.load_config()
     
     if args['drive']:
-        drive(cfg, model_path = args['--model'], use_joystick=args['--js'])
-
-    elif args['train']:
+        model_type = args['--type']
+        camera_type = args['--camera']
+        drive(cfg, model_path = args['--model'], use_joystick=args['--js'], model_type=model_type, camera_type=camera_type)
+    
+    if args['train']:
+        from train import multi_train
+        
         tub = args['--tub']
         model = args['--model']
-        cache = not args['--no_cache']
-        train(cfg, tub, model)
-
-
+        transfer = args['--transfer']
+        model_type = args['--type']
+        continuous = args['--continuous']
+        aug = args['--aug']     
+        multi_train(cfg, tub, model, transfer, model_type, continuous, aug)
